+++ src/types/cart.ts
@@ -1,6 +1,0 @@
-import { LineItem } from './line-item';
-
-export interface Cart {
-  lineItems: LineItem[];
-  subtotal: number;
-}