+++ src/types/line-item.ts
@@ -1,8 +1,0 @@
-import { Product } from './product';
-
-export interface LineItem {
-  id: string;
-  product: Product;
-  quantity: number;
-  price: number;
-}